gofmt; make it go gettable
