+++ etcdfs.go
@@ -1,23 +1,31 @@
 package main
 
 import (
-  "flag"
-  "log"
-  . "etcdfs"
-  "github.com/hanwen/go-fuse/fuse/pathfs"
-  "github.com/hanwen/go-fuse/fuse/nodefs"
+	"flag"
+	"fmt"
+	"github.com/hanwen/go-fuse/fuse/nodefs"
+	"github.com/hanwen/go-fuse/fuse/pathfs"
+	"github.com/xetorthio/etcd-fs/etcdfs"
+	"log"
+	"os"
 )
 
 func main() {
-  flag.Parse()
-  if len(flag.Args()) < 2 {
-    log.Fatal("Usage:\n  etcd-fs MOUNTPOINT ETCDENDPOINT")
-  }
-  etcdFs := EtcdFs{FileSystem: pathfs.NewDefaultFileSystem(), EtcdEndpoint: flag.Arg(1)}
-  nfs := pathfs.NewPathNodeFs(&etcdFs, nil)
-  server, _, err := nodefs.MountRoot(flag.Arg(0), nfs.Root(), nil)
-  if err != nil {
-    log.Fatalf("Mount fail: %v\n", err)
-  }
-  server.Serve()
+	flag.Usage = func() {
+		fmt.Fprintf(os.Stderr, "Usage:\n  etcd-fs MOUNTPOINT ETCDENDPOINT\n")
+	}
+	flag.Parse()
+	if len(flag.Args()) < 2 {
+		flag.Usage()
+	}
+	etcdFs := etcdfs.EtcdFs{
+		FileSystem:   pathfs.NewDefaultFileSystem(),
+		EtcdEndpoint: flag.Arg(1),
+	}
+	nfs := pathfs.NewPathNodeFs(&etcdFs, nil)
+	server, _, err := nodefs.MountRoot(flag.Arg(0), nfs.Root(), nil)
+	if err != nil {
+		log.Fatalf("Mount fail: %v\n", err)
+	}
+	server.Serve()
 }
